+++ src/client/services/orpc-client.ts
@@ -2,7 +2,7 @@ import { createORPCClient } from '@orpc/client'
 import type { ContractRouterClient } from '@orpc/contract'
 import type { JsonifiedClient } from '@orpc/openapi-client'
 import { OpenAPILink } from '@orpc/openapi-client/fetch'
-import { routerContract } from 'routes/otrpc.ts'
+import { routerContract } from '../../server/routes/otrpc-server.ts'
 import { port } from '../../env.ts'
 
 const link = new OpenAPILink(routerContract, {
